requirements.txt
celery==5.4.0
Django==4.2.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==3.13.0
django-filter==22.1
django_jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
flower==2.0.1
Pillow==9.2.0
pytz==2022.4
redis==5.0.4